M2.10c: Fix remaining Swift 6 compile errors
1. MetalKitSceneView: remove 'target: nil' from gesture recognizers
   (ambiguous with init(target:action:) Selector parameter)
2. SettingsView: replace deprecated 'placeholder:' with 'prompt:'
   (TextField API changed in iOS 17+)
3. SplatCacheManager: cast totalFileAllocatedSize to Int64 to match
   the accumulator type

+++ Model/SplatCacheManager.swift
@@ -80,7 +80,7 @@ actor SplatCacheManager {
         for case let url as URL in enumerator {
             let resources = try? url.resourceValues(forKeys: resourceKeys)
             if resources?.isRegularFile == true {
-                total += resources?.totalFileAllocatedSize ?? 0
+                total += Int64(resources?.totalFileAllocatedSize ?? 0)
             }
         }
         return total
